+++ src/components/MobileMenu.tsx
@@ -17,9 +17,9 @@ import { LibraryButton } from "./LibraryButton";
 import { PenModeButton } from "./PenModeButton";
 import { Stats } from "./Stats";
 import { actionToggleStats } from "../actions";
-import { mainMenuTunnel, welcomeScreenCenterTunnel } from "./LayerUI";
 import { HandButton } from "./HandButton";
 import { isHandToolActive } from "../appState";
+import { mainMenuTunnel, welcomeScreenCenterTunnel } from "./tunnels";
 
 type MobileMenuProps = {
   appState: AppState;
